refactor: rewrite cjsToEsm.cjs to ECMAScript, rename to esm.mjs
